Restart (slowly) on failure and poweroff system on multiple failures
Default systemd is to restart every 100ms on failure. This is suitable
for local services but when dealing with networks there is lag. With
the previous setup if the backend was unavailable for a moment then
the agent would fail several times and then hit the Burst limit in
the allowed Interval. Systemd then perm halts restarting the service.

Using this new config the service will restart every 10 seconds up to
10 times in a 120 second period. If the system still cannot recover
in that time then it will be powered off.